Silence compiler warning

+++ include/pybind11/eigen/tensor.h
@@ -138,7 +138,7 @@ struct get_tensor_descriptor {
 template <typename T, int size>
 std::vector<T> convert_dsizes_to_vector(const Eigen::DSizes<T, size> &arr) {
     std::vector<T> result(size);
-    for (size_t i = 0; i < size; i++) {
+    for (int i = 0; i < size; i++) {
         result[i] = arr[i];
     }
 
@@ -175,7 +175,7 @@ struct type_caster<Type, typename eigen_tensor_helper<Type>::ValidType> {
 
         Eigen::DSizes<typename Type::Index, Type::NumIndices> shape;
         for (size_t i = 0; i < Type::NumIndices; i++) {
-            shape[i] = arr.shape()[i];
+            shape[i] = arr.shape()[i ];
         }
 
         if (!Helper::is_correct_shape(shape)) {
